requirements.txt
numpy==1.23.5
pandas==1.5.3
pytorch==2.0.1
tqdm==4.65.0
scikit-learn==1.2.2
redis-py==4.3.4
fastparquet==2023.4.0
onnx==1.13.1
grpcio==1.42.0
flask==2.2.2
elasticsearch==8.8.2
tritonclient==2.34.0
requests==2.31.0